+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The current file describes significant changes in the code that may affect the u
 * Added `Oro\Bundle\AttachmentBundle\Provider\OriginalFileNameProvider` filename provider that
   uses a sanitized original filename for files if `attachment_original_filenames` feature is enabled.
 
+#### LocaleBundle
+* Added `oro:localization:localized-fallback-values:cleanup-unused` command that finds and deletes orphaned
+  `Oro\Bundle\LocaleBundle\Entity\LocalizedFallbackValue` entities that could appear due to disabled `orphanRemoval` option.
+* Added `cloneLocalizedFallbackValueAssociations()` method that is generated automatically and should be used in
+  `__clone()` for entities with localized fallback value relations to ensure correct cloning of localized fallback value
+  collections.
+
 ## 4.2.10
 
 ### Changed